+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.9.36] - 2022-06-08 11:41:00
+### Added
+- script preloader
+### Fixed
+- no-code editor configuration
+
 ## [4.9.35] - 2022-05-29 17:41:00
 ### Added
 - wishlist data